# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kutubxona",
        "description": "BLOCK jamoasi",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/book": {
            "get": {
                "tags": [
                    "Book API CRUD"
                ],
                "summary": "Get all Books",
                "description": "Retrieve all Books",
                "operationId": "book_index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book API CRUD"
                ],
                "summary": "Create a new Book",
                "description": "Add a new Book",
                "operationId": "book_store",
                "requestBody": {
                    "description": "Book save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
        "/api/v1/book/{id}": {
            "get": {
                "tags": [
                    "Book API CRUD"
                ],
                "summary": "Get a single Book by ID",
                "description": "Retrieve a single Book by its ID",
                "operationId": "book_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Book to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Book API CRUD"
                ],
                "summary": "Update a Book by ID",
                "description": "Update a specific Book by its ID",
                "operationId": "book_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Book to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Book save",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "status"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "_method": {
                                        "description": "Read-only: This property cannot be modified.",
                                        "type": "string",
                                        "example": "PUT"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Book API CRUD"
                ],
                "summary": "Delete a Book by ID",
                "description": "Remove a specific Book by its ID",
                "operationId": "book_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Book to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Book"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "title": "Book title",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Successfully"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "updated_at": {
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "ResponseError",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": []
                    },
                    "code": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}